requirements.txt
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
